stamp_catalog/requirements.txt
opencv-python
pillow
numpy
scikit-image
requests
flask